chapter_002/src/main/java/ru/job4j/condition/Point.java
package ru.job4j.condition;

import static java.lang.Math.pow;
import static java.lang.Math.sqrt;

/**
 * перегрузка конструкторов: расчет расстояния между двумя точками
 * в двухмерном и трехмерном пространствах
 * @author Evgeniy Kapaev
 * @version 1.1
 */
public class Point {
    private int x;
    private int y;
    private int z;

    public Point(int first, int second) {
        this.x = first;
        this.y = second;
    }

    public Point(int x, int y, int z) {
        this.x = x;
        this.y = y;
        this.z = z;
    }

    /**
     * расчет расстояния между точками в двухмерном пространстве
     */
    public double distance(Point that) {
        return sqrt(pow(this.x - that.x, 2) + pow(this.y - that.y, 2));
    }

    /**
     * расчет расстояния между точками в трехмерном пространстве
     */
    public double distance3d(Point that) {
        return sqrt(pow(this.x - that.x, 2) + pow(this.y - that.y, 2) + pow(this.z - that.z, 2));
    }
}
